+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.6.1
+### Bugfix
+* fix sourcemaps
+* fix prism + `--size`
+* fix echo double slashes
+* escape options to `$.terminal.format` (default to `true` - same behavior) to fix previous error
+
 ## 2.6.0
 ### Features
 * new option `mobileIngoreAutoSpace` to fix issue on Android when typing keys: `, ) .` (default empty array)